+++ wp-content/themes/virtual-hq-tour/src/js/map/assets/layers.png
