+++ wp-content/themes/virtual-hq-tour/src/js/map/assets/map--floor-2.webp
