+++ wp-content/themes/virtual-hq-tour/src/js/map/assets/marker-icon-2x.png
