+++ wp-content/themes/virtual-hq-tour/src/js/map/assets/marker-icon.png
