+++ wp-content/themes/virtual-hq-tour/src/js/map/classes/VideoModal.ts
@@ -1,14 +1,9 @@
-import { Modal } from 'bootstrap';
-import { VideoObject } from './VideoPopups';
+import BootstrapModal from '../../modules/BootstrapModal';
 
 /**
  * Dynamically create a Bootstrap Modal with a lite-vimeo web component
  */
-export class VideoModal {
-	private modal: Modal;
-	private modalTitle: HTMLElement;
-	private modalBody: HTMLElement;
-
+export class VideoModal extends BootstrapModal {
 	/**
 	 * Construct the Modal
 	 * @param trigger The button that triggers the modal
@@ -20,6 +15,8 @@ export class VideoModal {
 		modalTitleContent: string,
 		modalBodyContent: string
 	) {
+		const modalEl = document.getElementById( 'videoModal' )!;
+		super( modalEl );
 		this.modalTitle = document.getElementById( 'videoModalLabel' )!;
 		this.modalBody = document.getElementById( 'modal-video' )!;
 		const trigger = document.getElementById(
@@ -32,24 +29,13 @@ export class VideoModal {
 
 	/**
 	 * Construct the Modal's Contents
-	 * @param trigger The button that triggers the modal
+	 *
 	 * @param modalTitleContent The Modal Title
 	 * @param modalBodyContent The lite-vimeo web component(s)
 	 */
 	private initModal( modalTitleContent: string, modalBodyContent: string ) {
-		const modalEl = document.getElementById( 'videoModal' )!;
-		this.modal = new Modal( modalEl );
 		this.modalTitle.textContent = modalTitleContent;
 		this.modalBody.innerHTML = modalBodyContent;
 		this.modal.show();
-		modalEl.addEventListener(
-			'hidden.bs.modal',
-			this.resetModal.bind( this )
-		);
-	}
-
-	private resetModal() {
-		this.modalTitle.textContent = '';
-		this.modalBody.innerHTML = '';
 	}
 }
