+++ wp-content/themes/virtual-hq-tour/src/js/vendors/bootstrap.js
@@ -9,7 +9,7 @@ import '../../styles/vendors/bootstrap.scss';
 // import Carousel from 'bootstrap/js/dist/carousel';
 // import Collapse from 'bootstrap/js/dist/collapse';
 // import Dropdown from 'bootstrap/js/dist/dropdown';
-import Modal from 'bootstrap/js/dist/modal';
+// import Modal from 'bootstrap/js/dist/modal';
 import Offcanvas from 'bootstrap/js/dist/offcanvas';
 // import Popover from 'bootstrap/js/dist/popover';
 // import ScrollSpy from 'bootstrap/js/dist/scrollspy';
