+++ wp-content/themes/virtual-hq-tour/src/styles/abstracts/_variables.scss
@@ -16,7 +16,7 @@ $font-serif: adobe-garamond-pro, serif;
 
 // ABOUT
 $color-purple: #51276e;
-$color-purple--light: #7e42a8;
+$color-purple--light: #b6a0c6;
 $color-purple--dark: #330d48;
 
 // NEWS
