+++ wp-content/themes/virtual-hq-tour/style.css
@@ -4,7 +4,7 @@ Theme URI: https://github.com/choctaw-nation
 Author: Choctaw Nation of Oklahoma
 Author URI: https://www.choctawnation.com/
 Description: The Theme for the virtual hq tour of the Choctaw Nation of Oklahoma
-Version: 1.2.0
+Version: 1.3.0
 Requires at least: 6.0
 License: GPL-3.0-or-later
 License URI: https://www.gnu.org/licenses/gpl-3.0.html
